backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
bcrypt==4.0.1
pymongo
apscheduler
vaderSentiment
groq
aiosqlite
annotated-types
anyio
certifi
charset-normalizer
click
colorama
greenlet
h11
idna
pydantic
pydantic_core
requests
sniffio
starlette
typing-inspection
typing_extensions
urllib3